Add files via upload
First commit. Rabbit_feedback.py is able to connect to the scope and smaract, display the scope's screen, perform a rabbit scan, and analyze it to get the error signal. It cannot perform the feedback yet.
Rabbit_feedback is still a little redondant with Rabbit_scan: Rabbit_scan is imported to get the scanWidget only, the scan code is actually written in Rabbit_feedback.